backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
python-dotenv==1.0.1
imap-tools==1.6.0
aiosmtplib==3.0.2
python-keycloak==4.2.0
groq==0.11.0
openai==1.50.0
apscheduler==3.10.4
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-multipart==0.0.10
httpx<0.28.0
